# market-app/src/main/java/ru/yandex/practicum/my_market_app/model/dto/payment/ChargeStatus.java
/*
 * API Сервиса оплаты заказа
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


package ru.yandex.practicum.my_market_app.model.dto.payment;

import com.fasterxml.jackson.annotation.JsonProperty;
import lombok.AllArgsConstructor;
import lombok.EqualsAndHashCode;
import lombok.NoArgsConstructor;
import lombok.ToString;
import jakarta.validation.constraints.*;
import io.swagger.v3.oas.annotations.media.Schema;

/**
 * ChargeStatus
 */

@ToString
@EqualsAndHashCode
@AllArgsConstructor
@NoArgsConstructor
public class ChargeStatus {

    private String status;

    private Boolean isSuccess;

    /**
     * Get status
     *
     * @return status
     */
    @NotNull
    @Schema(name = "status", requiredMode = Schema.RequiredMode.REQUIRED)
    @JsonProperty("status")
    public String getStatus() {
        return status;
    }

    @JsonProperty("status")
    public void setStatus(String status) {
        this.status = status;
    }

    public ChargeStatus isSuccess(Boolean isSuccess) {
        this.isSuccess = isSuccess;
        return this;
    }

    /**
     * Get isSuccess
     *
     * @return isSuccess
     */
    @NotNull
    @Schema(name = "isSuccess", requiredMode = Schema.RequiredMode.REQUIRED)
    @JsonProperty("isSuccess")
    public Boolean getIsSuccess() {
        return isSuccess;
    }

    @JsonProperty("isSuccess")
    public void setIsSuccess(Boolean isSuccess) {
        this.isSuccess = isSuccess;
    }

}