# market-app/src/main/java/ru/yandex/practicum/my_market_app/model/dto/payment/Error.java
/*
 * API Сервиса оплаты заказа
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


package ru.yandex.practicum.my_market_app.model.dto.payment;
import com.fasterxml.jackson.annotation.JsonProperty;
import lombok.AllArgsConstructor;
import lombok.EqualsAndHashCode;
import lombok.NoArgsConstructor;
import lombok.ToString;
import io.swagger.v3.oas.annotations.media.Schema;


/**
 * Error
 */
@ToString
@EqualsAndHashCode
@AllArgsConstructor
@NoArgsConstructor
public class Error {

  private String message;

  /**
   * Get message
   *
   * @return message
   */

  @Schema(name = "message", example = "Сообщение об ошибке", requiredMode = Schema.RequiredMode.NOT_REQUIRED)
  @JsonProperty("message")
  public  String getMessage() {
    return message;
  }

  @JsonProperty("message")
  public void setMessage(String message) {
    this.message = message;
  }

}