# payment-service/src/main/java/ru/yandex/practicum/server/domain/ChargeStatus.java
package ru.yandex.practicum.server.domain;

import com.fasterxml.jackson.annotation.JsonProperty;
import lombok.AllArgsConstructor;
import lombok.EqualsAndHashCode;
import lombok.NoArgsConstructor;
import lombok.ToString;
import jakarta.validation.constraints.*;
import io.swagger.v3.oas.annotations.media.Schema;

/**
 * ChargeStatus
 */

@ToString
@EqualsAndHashCode
@AllArgsConstructor
@NoArgsConstructor
public class ChargeStatus {

    private String status;

    private Boolean isSuccess;

    /**
     * Get status
     *
     * @return status
     */
    @NotNull
    @Schema(name = "status", requiredMode = Schema.RequiredMode.REQUIRED)
    @JsonProperty("status")
    public String getStatus() {
        return status;
    }

    @JsonProperty("status")
    public void setStatus(String status) {
        this.status = status;
    }

    public ChargeStatus isSuccess(Boolean isSuccess) {
        this.isSuccess = isSuccess;
        return this;
    }

    /**
     * Get isSuccess
     *
     * @return isSuccess
     */
    @NotNull
    @Schema(name = "isSuccess", requiredMode = Schema.RequiredMode.REQUIRED)
    @JsonProperty("isSuccess")
    public Boolean getIsSuccess() {
        return isSuccess;
    }

    @JsonProperty("isSuccess")
    public void setIsSuccess(Boolean isSuccess) {
        this.isSuccess = isSuccess;
    }

}